// src/types/AxiosRequest.ts
export type AxiosMethod = 'get' | 'post' | 'put' | 'delete' | 'patch' | 'option'

export type AxiosQuery = {
  [props: string]: number | string | boolean
}

export type AxiosContentType =
  | 'application/x-www-form-urlencoded'
  | 'multipart/form-data'
  | 'application/json'
  | 'text/xml'

export type AxiosData = XMLHttpRequestBodyInit

export type AxiosHeader = {
  'Content-Type'?: AxiosContentType
  Range?: string
  Authroization?: string
}

export type AxiosReturnType = {
  status: 'success' | 'fail'
  data: JSON | ArrayBuffer | Blob | Document | string | { [props: string]: any }
}